refactor: Regroup isSnoozeAvailable and isReactionsAvailable in its own file
