+++ app/src/main/java/com/infomaniak/mail/utils/FeatureAvailability.kt
@@ -1,0 +1,37 @@
+/*
+ * Infomaniak Mail - Android
+ * Copyright (C) 2025 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.mail.utils
+
+import com.infomaniak.mail.data.LocalSettings
+import com.infomaniak.mail.data.LocalSettings.ThreadMode
+import com.infomaniak.mail.data.models.FeatureFlag
+import com.infomaniak.mail.data.models.mailbox.Mailbox
+
+object FeatureAvailability {
+    fun isSnoozeAvailable(featureFlags: Mailbox.FeatureFlagSet?, localSettings: LocalSettings): Boolean {
+        fun hasSnoozeFeatureFlag() = featureFlags?.contains(FeatureFlag.SNOOZE) == true
+        fun isConversationMode() = localSettings.threadMode == ThreadMode.CONVERSATION
+        return hasSnoozeFeatureFlag() && isConversationMode()
+    }
+
+    fun isReactionsAvailable(featureFlags: Mailbox.FeatureFlagSet?, localSettings: LocalSettings): Boolean {
+        fun hasEmojiFeatureFlag() = featureFlags?.contains(FeatureFlag.EMOJI_REACTION) == true
+        fun isConversationMode() = localSettings.threadMode == ThreadMode.CONVERSATION
+        return hasEmojiFeatureFlag() && isConversationMode()
+    }
+}
